+++ src/MQTTProtocolOut.c
@@ -45,7 +45,7 @@ char* MQTTProtocol_addressPort(const char* uri, int* port)
 {
 	char* colon_pos = strrchr(uri, ':'); /* reverse find to allow for ':' in IPv6 addresses */
 	char* buf = (char*)uri;
-	int len;
+	size_t len;
 
 	FUNC_ENTRY;
 	if (uri[0] == '[')
@@ -56,7 +56,7 @@ char* MQTTProtocol_addressPort(const char* uri, int* port)
 
 	if (colon_pos)
 	{
-		int addr_len = colon_pos - uri;
+		size_t addr_len = colon_pos - uri;
 		buf = malloc(addr_len + 1);
 		*port = atoi(colon_pos + 1);
 		MQTTStrncpy(buf, uri, addr_len+1);
